* commit 'f9f69089694af42f01e3ad330c871fc316ae8efa': (31 commits)
  Fix table list context and gear menus incorrectly hiding separator.
  Add missing synthesize.
  Reorganise some logical file groupings.
  #2978: Fix direct private ivar access and add warnings to more occurrences.
  Fix dependency for parallel builds.
  #3321: Fix more search search paths and deprecation warnings.
  #3221: Upgrade QueryKit project settings to silence warnings.
  Fix command line builds and remove incorrect framework paths.
  Update framework readme's.
  Update issue templates
  #3213: Switch to using SPMainQSync.
  Prevent users from creating tables using the performance_schema storage engine.
  Fix type warning.
  Fix nullability specifiers.
  Remove unnecessary import.
  Move issue templates to correct directory.
  Add GitHub issue templates.
  #3213: Fix crash caused by c37bea8.
  Fix UI access from background query execution thread.
  Remove QueryKit scheme to ensure the Sequel Pro one is selected by default upon Xcode launch.
  ...

+++ Source/SPWindowController.m
@@ -239,6 +239,7 @@
 	[newWindow setTitle:[[[[tabView selectedTabViewItem] identifier] parentWindow] title]];
 
 	// New window's tabBar control
+#warning Private ivar accessed from outside (#2978)
 	PSMTabBarControl *control = [newWindowController valueForKey:@"tabBar"];
 
 	// Add the selected tab to the new window
@@ -849,6 +850,7 @@
 	[newWindow setTitle:[[[tabViewItem identifier] parentWindow] title]];
 
 	// Return the window's tab bar
+#warning Private ivar accessed from outside (#2978)
 	return [newWindowController valueForKey:@"tabBar"];
 }
 
@@ -856,7 +858,7 @@
  * When dragging a tab off the tab bar, return an image so that a
  * drag placeholder can be displayed.
  */
-- (NSImage *)tabView:(NSTabView *)aTabView imageForTabViewItem:(NSTabViewItem *)tabViewItem offset:(NSSize *)offset styleMask:(unsigned int *)styleMask
+- (NSImage *)tabView:(NSTabView *)aTabView imageForTabViewItem:(NSTabViewItem *)tabViewItem offset:(NSSize *)offset styleMask:(NSUInteger *)styleMask
 {
 	NSImage *viewImage = [[NSImage alloc] init];
